feat: implement contact panel functionality with open/close logic and event listeners
